Speed up getDelta() by using explicit inverse
Previously, we computed ||H*x|| by evaluating
    Hx = Aty_r / H
which uses the MATLAB backslash operator. This is the conventional wisdom
since H is an upper triangular matrix and this linear solve (via the BLAS
TRSV method) requires fewer operations than a general matrix multiplication
(BLAS GEMM), not to mention the fact that we never have to explicitly
compute the inverse of H.

However, gpuArray GEMM is ridiculously fast, way faster than TRSV for
matrices of these sizes, and so I have switched this to:
    Hx = Aty_r * inv(H)
We can cache the value of inv(H), so the cost of computing this inverse is
negligible compared to the cost of applying it, and this problem is usually
well-conditioned enough that the improved numerics of TRSV don't matter.