Correct the BasisOptIncoh penalty term
Previously, the BasisOptIncoh penalty term was not correctly defined in
cases where the subsample interpolation ratio R > 1. It assumed that the
subsample-shifted versions of the basis would still be approximately
orthonormal. This is a reasonable assumption when the coherence penalty
weight is small, but as we increase this weight, it creates an incentive
to produce bases that are rank-deficient after being subsample-shifted,
which is undesirable.

This commit redefines the penalty term to properly handle the R > 1 case
(while remaining identical to before in the R = 1 case). Testing shows
that this produces more reasonable results, even with a large coherence
penalty weight.

However, this change will require recomputation of any bases that were
fitted using the older version of this code.