Don't check error identifier on Math.has_gpu()
The has_gpu() function determines whether a GPU is available by trying
to construct a gpuArray and seeing if this produces an error.
Previously, this checked whether the error identifier matched a
particular string. However, it turns out that different versions of
MATLAB produce different error identifiers for the same error. This
commit relaxes this error checking, so any error encountered while
trying to perform `x = gpuArray();` is considered to be evidence that
a GPU is not available on this machine.