Bugfix (avoid calling quantile() on gpuArrays)
It appears that quantile() can produce a fairly opaque error:
    "radix_sort: failed to get memory buffer"
when called on large gpuArrays. This error is particularly annoying
because it has no stack trace (making it hard to debug) and is not
caught by a try/catch block. So I am just going to move it to host
memory and do it on CPU.